refactor(cli): extract docs command module
